Reverted to 1d64f06a, but keep minor optimization (RemainingDepth).

+++ Serialization/NodeDeserializer.cs
@@ -48,7 +48,7 @@ internal sealed class NodeDeserializer : JsonConverter<DeserializedNode>
                             break;
                         case _R:
                             reader.Read();
-                            node.RemainingDepth = reader.GetInt16();
+                            node.RemainingDepth = reader.GetByte();
                             break;
                         case _W:
                             reader.Read();
